+++ src/main/java/com/multipjt/multi_pjt/Badge/service/IBadgeService.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public interface IBadgeService {
     // 특정 회원의 현재 점수를 조회하는 메서드
-    BigDecimal getCurrentPoints(Long memberId);
+    BigDecimal getCurrentPoints(int memberId);
     
     // 주어진 점수에 따라 회원의 뱃지 레벨을 결정하는 메서드
     String getBadgeLevel(BigDecimal currentPoints);
@@ -15,14 +15,14 @@ public interface IBadgeService {
     BigDecimal getPointsToNextBadge(BigDecimal currentPoints);
     
     // 특정 회원의 점수를 업데이트하는 메서드
-    void updateUserPoints(Long memberId, BigDecimal points);
+    void updateUserPoints(int memberId, BigDecimal points);
     
     // 모든 회원의 점수 순위를 조회하는 메서드
     List<Map<String, Object>> getUserRankings();
     
     // 특정 회원의 모든 활동을 처리하는 메서드
-    void processUserActivities(Long memberId);
+    void processUserActivities(int memberId);
     
     // 특정 회원의 뱃지를 생성하는 메서드
-    void createBadge(Long memberId);
+    void createBadge(int memberId);
 }
